[AIRFLOW-6391] Move content of utils.tests to tests.test_utils (#6949)
